Factor out CTA for missing time dimension
