+++ web-common/src/features/dashboards/time-controls/TimeControls.svelte
@@ -8,12 +8,6 @@ Constructs a TimeSeriesTimeRange object – to be used as the filter in MetricsE
 We should rename TimeSeriesTimeRange to a better name.
 -->
 <script lang="ts">
-  import { goto } from "$app/navigation";
-  import Calendar from "@rilldata/web-common/components/icons/Calendar.svelte";
-  import Shortcut from "@rilldata/web-common/components/tooltip/Shortcut.svelte";
-  import Tooltip from "@rilldata/web-common/components/tooltip/Tooltip.svelte";
-  import TooltipContent from "@rilldata/web-common/components/tooltip/TooltipContent.svelte";
-  import TooltipShortcutContainer from "@rilldata/web-common/components/tooltip/TooltipShortcutContainer.svelte";
   import {
     useModelAllTimeRange,
     useModelHasTimeSeries,
@@ -27,8 +21,8 @@ We should rename TimeSeriesTimeRange to a better name.
   import { useRuntimeServiceGetCatalogEntry } from "@rilldata/web-common/runtime-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import type { UseQueryStoreResult } from "@sveltestack/svelte-query";
-  import { selectTimestampColumnFromSchema } from "../../metrics-views/column-selectors";
   import { metricsExplorerStore, useDashboardStore } from "../dashboard-stores";
+  import NoTimeDimensionCTA from "./NoTimeDimensionCTA.svelte";
   import {
     addGrains,
     checkValidTimeGrain,
@@ -88,25 +82,6 @@ We should rename TimeSeriesTimeRange to a better name.
     makeTimeSeriesTimeRangeAndUpdateAppState(baseTimeRange, timeGrain);
   }
 
-  let timestampColumns: Array<string>;
-
-  let modelQuery;
-  $: if ($metricsViewQuery?.data?.entry?.metricsView?.model) {
-    modelQuery = useRuntimeServiceGetCatalogEntry(
-      $runtimeStore.instanceId,
-      $metricsViewQuery?.data?.entry?.metricsView?.model
-    );
-  }
-
-  $: if ($modelQuery && $modelQuery.isSuccess && !$modelQuery.isRefetching) {
-    const model = $modelQuery.data?.entry?.model;
-    timestampColumns = selectTimestampColumnFromSchema(model?.schema);
-  } else {
-    timestampColumns = [];
-  }
-
-  $: redirectToScreen = timestampColumns?.length > 0 ? "metrics" : "model";
-
   // we get the timeGrainOptions so that we can assess whether or not the
   // activeTimeGrain is valid whenever the baseTimeRange changes
   let timeGrainOptions: TimeGrainOption[];
@@ -164,35 +139,14 @@ We should rename TimeSeriesTimeRange to a better name.
 
     metricsExplorerStore.setSelectedTimeRange(metricViewName, newTimeRange);
   }
-
-  function noTimeseriesCTA() {
-    if (timestampColumns?.length) {
-      goto(`/dashboard/${metricViewName}/edit`);
-    } else {
-      const modelName = $metricsViewQuery?.data?.entry?.metricsView?.model;
-      goto(`/model/${modelName}`);
-    }
-  }
 </script>
 
 <div class="flex flex-row">
   {#if !hasTimeSeries}
-    <Tooltip location="bottom" distance={8}>
-      <div
-        on:click={() => noTimeseriesCTA()}
-        class="px-3 py-2 flex flex-row items-center gap-x-3 cursor-pointer"
-      >
-        <span class="ui-copy-icon"><Calendar size="16px" /></span>
-        <span class="ui-copy-disabled">No time dimension specified</span>
-      </div>
-      <TooltipContent slot="tooltip-content" maxWidth="250px">
-        Add a time dimension to your {redirectToScreen} to enable time series plots.
-        <TooltipShortcutContainer>
-          <div class="capitalize">Edit {redirectToScreen}</div>
-          <Shortcut>Click</Shortcut>
-        </TooltipShortcutContainer>
-      </TooltipContent>
-    </Tooltip>
+    <NoTimeDimensionCTA
+      {metricViewName}
+      modelName={$metricsViewQuery?.data?.entry?.metricsView?.model}
+    />
   {:else}
     <TimeRangeSelector
       {metricViewName}
